feat(websocket): Refactor WebSocket module for improved structure
Refactor the WebSocket infrastructure by moving and reorganizing related files into a dedicated 'websocket' package. This includes:

- Moving DTOs, publishers, controllers, authentication, and presence listeners to new 'org.ping_me.websocket' subpackages.
- Deleting old WebSocket-related files from 'org.ping_me.config', 'org.ping_me.dto
