--- src/main/java/org/ping_me/dto/ws/chat/common/BaseChatEventPayload.java
@@ -1,18 +1,0 @@
-package org.ping_me.dto.ws.chat.common;
-
-import lombok.Getter;
-
-/**
- * Admin 11/20/2025
- *
- **/
-
-public abstract class BaseChatEventPayload {
-
-    @Getter
-    private final ChatEventType chatEventType;
-
-    public BaseChatEventPayload(ChatEventType chatEventType) {
-        this.chatEventType = chatEventType;
-    }
-}
