--- src/main/java/org/ping_me/dto/ws/chat/room/RoomMemberAddedEventPayload.java
@@ -1,34 +1,0 @@
-package org.ping_me.dto.ws.chat.room;
-
-import lombok.Getter;
-import org.ping_me.dto.response.chat.message.MessageResponse;
-import org.ping_me.dto.response.chat.room.RoomResponse;
-import org.ping_me.dto.ws.chat.common.BaseChatEventPayload;
-import org.ping_me.dto.ws.chat.common.ChatEventType;
-
-/**
- * Admin 11/20/2025
- *
- **/
-@Getter
-public class RoomMemberAddedEventPayload extends BaseChatEventPayload {
-
-    private final RoomResponse roomResponse;
-    private final Long targetUserId;
-    private final Long actorUserId;
-    private MessageResponse systemMessage;
-
-    public RoomMemberAddedEventPayload(
-            RoomResponse dto,
-            Long targetUserId,
-            Long actorUserId,
-            MessageResponse systemMessage
-    ) {
-        super(ChatEventType.MEMBER_ADDED);
-        this.roomResponse = dto;
-        this.targetUserId = targetUserId;
-        this.actorUserId = actorUserId;
-        this.systemMessage = systemMessage;
-    }
-
-}
