--- src/main/java/org/ping_me/dto/ws/friendship/common/FriendshipEventType.java
@@ -1,5 +1,0 @@
-package org.ping_me.dto.ws.friendship.common;
-
-public enum FriendshipEventType {
-    INVITED, ACCEPTED, REJECTED, CANCELED, DELETED
-}
