+++ src/main/java/org/ping_me/websocket/config/WebSocketConfiguration.java
@@ -1,0 +1,56 @@
+package org.ping_me.websocket.config;
+
+import lombok.RequiredArgsConstructor;
+import org.ping_me.websocket.auth.StompAuthInterceptor;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.messaging.simp.config.ChannelRegistration;
+import org.springframework.messaging.simp.config.MessageBrokerRegistry;
+import org.springframework.scheduling.concurrent.ThreadPoolTaskScheduler;
+import org.springframework.web.socket.config.annotation.EnableWebSocketMessageBroker;
+import org.springframework.web.socket.config.annotation.StompEndpointRegistry;
+import org.springframework.web.socket.config.annotation.WebSocketMessageBrokerConfigurer;
+
+/**
+ * Admin 8/10/2025
+ **/
+@Configuration
+@EnableWebSocketMessageBroker
+@RequiredArgsConstructor
+public class WebSocketConfiguration implements WebSocketMessageBrokerConfigurer {
+
+    private final StompAuthInterceptor stompAuthInterceptor;
+
+    @Value("${app.websocket.heartbeat.server-send-interval}")
+    private long serverSendInterval;
+
+    @Value("${app.websocket.heartbeat.server-receive-interval}")
+    private long serverReceiveInterval;
+
+
+    @Override
+    public void registerStompEndpoints(StompEndpointRegistry registry) {
+        registry
+                .addEndpoint("/core-service/ws")
+                .setAllowedOriginPatterns("*")
+                .withSockJS();
+    }
+
+    @Override
+    public void configureMessageBroker(MessageBrokerRegistry registry) {
+        var scheduler = new ThreadPoolTaskScheduler();
+        scheduler.setPoolSize(4);
+        scheduler.setThreadNamePrefix("pingme-wss-");
+        scheduler.initialize();
+        registry
+                .enableSimpleBroker("/topic", "/queue")
+                .setHeartbeatValue(new long[]{serverSendInterval, serverReceiveInterval})
+                .setTaskScheduler(scheduler);
+        registry.setUserDestinationPrefix("/user");
+    }
+
+    @Override
+    public void configureClientInboundChannel(ChannelRegistration registration) {
+        registration.interceptors(stompAuthInterceptor);
+    }
+}
