+++ src/main/java/org/ping_me/websocket/dto/WsBroadcastWrapper.java
@@ -1,0 +1,22 @@
+package org.ping_me.websocket.dto;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.experimental.FieldDefaults;
+
+/**
+ * Admin 2/17/2026
+ *
+ **/
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@FieldDefaults(level = AccessLevel.PRIVATE)
+public class WsBroadcastWrapper {
+
+    String destination;
+    Object payload;
+
+}
