+++ src/main/java/org/ping_me/websocket/dto/chat/common/BaseChatEventPayload.java
@@ -1,0 +1,18 @@
+package org.ping_me.websocket.dto.chat.common;
+
+import lombok.Getter;
+
+/**
+ * Admin 11/20/2025
+ *
+ **/
+
+public abstract class BaseChatEventPayload {
+
+    @Getter
+    private final ChatEventType chatEventType;
+
+    public BaseChatEventPayload(ChatEventType chatEventType) {
+        this.chatEventType = chatEventType;
+    }
+}
