+++ src/main/java/org/ping_me/websocket/dto/chat/room/RoomUpdatedEventPayload.java
@@ -1,0 +1,24 @@
+package org.ping_me.websocket.dto.chat.room;
+
+import lombok.Getter;
+import org.ping_me.dto.response.chat.message.MessageResponse;
+import org.ping_me.dto.response.chat.room.RoomResponse;
+import org.ping_me.websocket.dto.chat.common.BaseChatEventPayload;
+import org.ping_me.websocket.dto.chat.common.ChatEventType;
+
+/**
+ * Admin 8/30/2025
+ *
+ **/
+@Getter
+public class RoomUpdatedEventPayload extends BaseChatEventPayload {
+
+    private final RoomResponse roomResponse;
+    private final MessageResponse systemMessage;
+
+    public RoomUpdatedEventPayload(RoomResponse roomResponse, MessageResponse systemMessage) {
+        super(ChatEventType.ROOM_UPDATED);
+        this.roomResponse = roomResponse;
+        this.systemMessage = systemMessage;
+    }
+}
