+++ src/main/java/org/ping_me/websocket/dto/friendship/common/FriendshipEventType.java
@@ -1,0 +1,5 @@
+package org.ping_me.websocket.dto.friendship.common;
+
+public enum FriendshipEventType {
+    INVITED, ACCEPTED, REJECTED, CANCELED, DELETED
+}
