+++ src/main/java/org/ping_me/websocket/presence/UserPresenceListener.java
@@ -1,0 +1,86 @@
+package org.ping_me.websocket.presence;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.ping_me.websocket.WebSocketDestinations;
+import org.ping_me.websocket.auth.UserSocketPrincipal;
+import org.ping_me.websocket.dto.user_status.UserOnlineStatusRespone;
+import org.ping_me.model.chat.Friendship;
+import org.ping_me.model.constant.UserStatus;
+import org.ping_me.service.friendship.FriendshipService;
+import org.ping_me.service.user.CurrentUserProfileService;
+import org.ping_me.utils.mapper.UserMapper;
+import org.springframework.context.event.EventListener;
+import org.springframework.messaging.simp.SimpMessagingTemplate;
+import org.springframework.messaging.simp.stomp.StompHeaderAccessor;
+import org.springframework.stereotype.Component;
+import org.springframework.web.socket.messaging.SessionConnectedEvent;
+import org.springframework.web.socket.messaging.SessionDisconnectEvent;
+
+@Component
+@RequiredArgsConstructor
+@Slf4j
+public class UserPresenceListener {
+
+    // Websocket
+    private final SimpMessagingTemplate messagingTemplate;
+
+    // Service
+    private final CurrentUserProfileService currentUserProfileService;
+    private final FriendshipService friendshipService;
+
+    // Mapper
+    private final UserMapper userMapper;
+
+    @EventListener
+    public void handleConnectEvent(SessionConnectedEvent event) {
+        StompHeaderAccessor accessor = StompHeaderAccessor.wrap(event.getMessage());
+        UserSocketPrincipal userPrincipal = userMapper.extractUserPrincipal(accessor.getUser());
+
+        if (userPrincipal == null) {
+            return;
+        }
+
+        Long userId = userPrincipal.getId();
+        String name = userPrincipal.getName();
+
+        currentUserProfileService.updateStatus(userId, UserStatus.ONLINE);
+        var friends = friendshipService.getAllFriendshipsOfCurrentUser(userPrincipal.getEmail());
+        var payload = new UserOnlineStatusRespone(userId, name, true);
+
+        for (Friendship f : friends) {
+            Long friendId = (f.getUserA().getId().equals(userId)) ? f.getUserB().getId() : f.getUserA().getId();
+            messagingTemplate.convertAndSendToUser(
+                    String.valueOf(friendId),
+                    WebSocketDestinations.USER_QUEUE_STATUS_SUFFIX,
+                    payload
+            );
+        }
+    }
+
+    @EventListener
+    public void handleDisconnectEvent(SessionDisconnectEvent event) {
+        StompHeaderAccessor accessor = StompHeaderAccessor.wrap(event.getMessage());
+        UserSocketPrincipal userPrincipal = userMapper.extractUserPrincipal(accessor.getUser());
+
+        if (userPrincipal == null) {
+            return;
+        }
+
+        Long userId = userPrincipal.getId();
+        String name = userPrincipal.getName();
+
+        currentUserProfileService.updateStatus(userId, UserStatus.OFFLINE);
+        var friends = friendshipService.getAllFriendshipsOfCurrentUser(userPrincipal.getEmail());
+        var payload = new UserOnlineStatusRespone(userId, name, false);
+
+        for (Friendship f : friends) {
+            Long friendId = (f.getUserA().getId().equals(userId)) ? f.getUserB().getId() : f.getUserA().getId();
+            messagingTemplate.convertAndSendToUser(
+                    String.valueOf(friendId),
+                    WebSocketDestinations.USER_QUEUE_STATUS_SUFFIX,
+                    payload
+            );
+        }
+    }
+}
