+++ src/main/java/org/ping_me/websocket/publisher/FriendshipEventPublisher.java
@@ -1,0 +1,75 @@
+package org.ping_me.websocket.publisher;
+
+import org.ping_me.dto.response.user.UserSummaryResponse;
+import lombok.RequiredArgsConstructor;
+import org.ping_me.websocket.WebSocketDestinations;
+import org.ping_me.websocket.dto.friendship.FriendshipEventPayload;
+import org.ping_me.websocket.dto.friendship.common.FriendshipEventType;
+import org.ping_me.model.User;
+import org.ping_me.model.chat.Friendship;
+import org.ping_me.service.friendship.event.*;
+import org.springframework.stereotype.Component;
+import org.springframework.transaction.event.TransactionPhase;
+import org.springframework.transaction.event.TransactionalEventListener;
+
+@Component
+@RequiredArgsConstructor
+public class FriendshipEventPublisher {
+
+    private final RedisWsPublisher redisWsPublisher;
+
+    @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
+    public void handleFriendshipInvited(FriendshipInvitedEvent event) {
+        var friendship = event.getFriendship();
+
+        // Gửi thông báo cho cả người gửi và người nhận lời mời
+        sendViaRedis(friendship.getUserA().getId(), FriendshipEventType.INVITED, friendship);
+        sendViaRedis(friendship.getUserB().getId(), FriendshipEventType.INVITED, friendship);
+    }
+
+    @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
+    public void handleFriendshipAccepted(FriendshipAcceptedEvent event) {
+        sendViaRedis(event.getTargetId(), FriendshipEventType.ACCEPTED, event.getFriendship());
+    }
+
+    @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
+    public void handleFriendshipRejected(FriendshipRejectedEvent event) {
+        sendViaRedis(event.getTargetId(), FriendshipEventType.REJECTED, event.getFriendship());
+    }
+
+    @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
+    public void handleFriendshipCanceled(FriendshipCanceledEvent event) {
+        sendViaRedis(event.getTargetId(), FriendshipEventType.CANCELED, event.getFriendship());
+    }
+
+    @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
+    public void handleFriendshipDeleted(FriendshipDeletedEvent event) {
+        sendViaRedis(event.getTargetId(), FriendshipEventType.DELETED, event.getFriendship());
+    }
+
+    private void sendViaRedis(Long targetId, FriendshipEventType type, Friendship friendship) {
+        var payload = buildPayload(type, friendship, targetId);
+        redisWsPublisher.publish(WebSocketDestinations.userQueueFriendship(targetId), payload);
+    }
+
+    private FriendshipEventPayload buildPayload(FriendshipEventType type, Friendship friendship, Long targetId) {
+        User user = friendship.getUserA().getId().equals(targetId)
+                ? friendship.getUserB()
+                : friendship.getUserA();
+
+        var userSummaryResponse = new UserSummaryResponse(
+                user.getId(),
+                user.getEmail(),
+                user.getName(),
+                user.getAvatarUrl(),
+                user.getStatus(),
+                null
+        );
+        userSummaryResponse.setFriendshipSummary(new UserSummaryResponse.FriendshipSummary(
+                friendship.getId(),
+                friendship.getFriendshipStatus()
+        ));
+
+        return new FriendshipEventPayload(type, userSummaryResponse);
+    }
+}
